Require environment to be passed in from command line.
It sucks that this extra_arguments stanza needs to be repeated in each
module, but we're only allowed to define one terraform {} block and
modules need to define their source.

The good news is that this strategy lets us define all the environment
references (both terragrunt locking/state and terraform code
remote_state) in one logical place, the terragrunt {} block in the
terraform.tfvars assembled after inclusion.