Fix a regex bug.
I tried to write a test to trigger the behavior I saw when actually
running kitchen, but I couldn't repro the problem even with the exact
args provided to the wrapper by kitchen. So I gave up.

Behavior was: '-var-file=/path/a -destroy=true /path/b' was munged to
'-var-file=/path/a /b'.